Crate vexyz_math

Macros

bvec2!
Builder macro for creating new 2-dimensional vectors with boolean components.

bvec3!
Builder macro for creating new 3-dimensional vectors with boolean components.

bvec4!
Builder macro for creating new 4-dimensional vectors with boolean components.

ivec2!
Builder macro for creating new 2-dimensional vectors with integer components.

ivec3!
Builder macro for creating new 3-dimensional vectors with integer components.

ivec4!
Builder macro for creating new 4-dimensional vectors with integer components.

mat2!
Builder macro for creating new 2x2 column major matrices.

mat3!
Builder macro for creating new 3x3 column major matrices.

mat4!
Builder macro for creating new 4x4 column major matrices.

quat!
Builder macro for creating new quaternions.

vec2!
Builder macro for creating new 2-dimensional vectors with floating point components.

vec3!
Builder macro for creating new 3-dimensional vectors with floating point components.

vec4!
Builder macro for creating new 4-dimensional vectors with floating point components.

Structs

Mat2
Column major matrix with 2 columns and 2 rows.

Mat3
Column major matrix with 3 columns and 3 rows.

Mat4
Column major matrix with 4 columns and 4 rows.

Quat
Quaternion with 4 floating-point components a, b, c, and d. Unit quaternions represent rotation.

Vec2
2-dimensional vector with floating point x, and y components.

Vec2b
2-dimensional vector with boolean x, and y components.

Vec2i
2-dimensional vector with integer x, and y components.

Vec3
3-dimensional vector with floating point x, y, and z components.

Vec3b
3-dimensional vector with boolean x, y, and z components.

Vec3i
3-dimensional vector with integer x, y, and z components.

Vec4
4-dimensional vector with floating point x, y, z, and w components.

Vec4b
4-dimensional vector with boolean x, y, z, and w components.

Vec4i
4-dimensional vector with integer x, y, z, and w components.

Traits

FloatOps

Mat2Mat2Ops

Mat3Mat3Ops

Mat4Mat4Ops

QuatQuatOps

QuatVec3Ops

Vec2Ops

Vec2iOps

Vec3Ops

Vec3iOps

Vec4Ops

Vec4iOps