Struct vexyz_math::Quat

pub struct Quat {
    // some fields omitted
}

Quaternion with 4 floating-point components a, b, c, and d. Unit quaternions represent rotation.

Multiple rotations are combine using quaternion multiplication. For example, given rotations r1 and r2, you can obtain a combined rotation that performs r1 then r2 as r2*r1. Note the order of arguments! Quaternion multiplication is not associative, so r2*r1 != r1*r2.

Alternatively, you can use rotate_q() method: let r3 = r1.rotate_q(r2). Using rotate() method is more intuitive, because the resulting transformation is equivalent to applying all the operations one at a time, from left to right.

When combining multiple rotations, the resulting quaternion will accumulate floating point errors. You can fix that by periodically normalizing the quaterion with q.normalize().

Aside from rotation and multiplication, Vexyz provides a variety of component-wise quaternion operators and methods.

Methods

impl Quat

fn new(a: f64, b: f64, c: f64, d: f64) -> Self

Constructs a new Quat.

fn a(&self) -> f64

Component accessor, returns the 1st component of the quaternion.

fn b(&self) -> f64

Component accessor, returns the 2nd component of the quaternion.

fn c(&self) -> f64

Component accessor, returns the 3rd component of the quaternion.

fn d(&self) -> f64

Component accessor, returns the 4th component of the quaternion.

fn sum(&self) -> f64

Returns the sum of vector components.

Examples

#[macro_use] extern crate vexyz_math;
use vexyz_math::*;

let u = quat!(20, 30, 40, 50);
assert_eq!(u.sum(), 20.0 + 30.0 + 40.0 + 50.0);

fn abs(&self) -> Quat

Performs abs() on each component, producing a new vector.

Examples

#[macro_use] extern crate vexyz_math;
use vexyz_math::*;

let u = quat!(20, -30, 40, -50);
assert_eq!(u.abs(), quat!(20, 30, 40, 50));

fn rotate_x(&self, rads: f64) -> Quat

Returns combined rotation of the lhs quaternion followed by rotation around x axis.

Examples

#[macro_use] extern crate vexyz_math;
use vexyz_math::*;

//XXX add test

fn rotate_y(&self, rads: f64) -> Quat

Returns combined rotation of the lhs quaternion followed by rotation around y axis.

Examples

#[macro_use] extern crate vexyz_math;
use vexyz_math::*;

//XXX add test

fn rotate_z(&self, rads: f64) -> Quat

Returns combined rotation of the lhs quaternion followed by rotation around z axis.

Examples

#[macro_use] extern crate vexyz_math;
use vexyz_math::*;

//XXX add test

fn as_vec4(&self) -> Vec4

Extracts quaternion components into a vector: Quat(a, b, c, d) -> Vec4(b, c, d, a).

Examples

#[macro_use] extern crate vexyz_math;
use vexyz_math::*;

let q = quat!(2, 3, 4, 5);
assert_eq!(q.as_vec4(), vec4!(3, 4, 5, 2));

fn length(&self) -> f64

Computes the norm of the quaternion (similar to vector length).

Examples

#[macro_use] extern crate vexyz_math;
use vexyz_math::*;

let u = quat!(20, 30, 40, 50);
assert_eq!(u.length(), ((20*20 + 30*30 + 40*40 + 50*50) as f64).sqrt());

fn normalize(&self) -> Quat

Normalizes the quaternion.

Examples

#[macro_use] extern crate vexyz_math;
use vexyz_math::*;

let q = quat!(20, 30, 40, 50);
assert!(q.normalize().norm().approx_equal(1.0, 1e-8));