Struct vexyz_math::Vec2i

pub struct Vec2i {
    // some fields omitted
}

2-dimensional vector with integer x, and y components.

Vectors can also represent colors with the help of r(), and g() accessors.

Most operators and methods on vectors are performed in component-wise fashion. The notable exception is vector-matrix multiplication.

Methods

impl Vec2i

fn new(x: i32, y: i32) -> Self

Constructs a new Vec2i.

fn x(&self) -> i32

Component accessor, returns the 1st component of the vector.

fn y(&self) -> i32

Component accessor, returns the 2nd component of the vector.

fn r(&self) -> i32

Color-style component accessor, returns the 1st component of the vector.

fn g(&self) -> i32

Color-style component accessor, returns the 2nd component of the vector.

fn sum(&self) -> i32

Returns the sum of vector components.

Examples

#[macro_use] extern crate vexyz_math;
use vexyz_math::*;

let u = ivec2!(20, 30);
assert_eq!(u.sum(), 20 + 30);

fn abs(&self) -> Vec2i

Performs abs() on each component, producing a new vector.

Examples

#[macro_use] extern crate vexyz_math;
use vexyz_math::*;

let u = ivec2!(20, -30);
assert_eq!(u.abs(), ivec2!(20, 30));