Struct vexyz_math::Vec4b

pub struct Vec4b {
    // some fields omitted
}

4-dimensional vector with boolean x, y, z, and w components.

Vectors can also represent colors with the help of r(), g(), b(), and a() accessors.

Most operators and methods on vectors are performed in component-wise fashion. The notable exception is vector-matrix multiplication.

Methods

impl Vec4b

fn new(x: bool, y: bool, z: bool, w: bool) -> Self

Constructs a new Vec4b.

fn x(&self) -> bool

Component accessor, returns the 1st component of the vector.

fn y(&self) -> bool

Component accessor, returns the 2nd component of the vector.

fn z(&self) -> bool

Component accessor, returns the 3rd component of the vector.

fn w(&self) -> bool

Component accessor, returns the 4th component of the vector.

fn r(&self) -> bool

Color-style component accessor, returns the 1st component of the vector.

fn g(&self) -> bool

Color-style component accessor, returns the 2nd component of the vector.

fn b(&self) -> bool

Color-style component accessor, returns the 3rd component of the vector.

fn a(&self) -> bool

Color-style component accessor, returns the 4th component of the vector.

fn all(&self) -> bool

Returns true if all components of the vector are true, false otherwise.

Examples

#[macro_use] extern crate vexyz_math;
use vexyz_math::*;

assert!(bvec4!(true, true, true, true).all());
assert!(!bvec4!(true, false, true, true).all());

fn any(&self) -> bool

Returns true if at least one of the vector components is true, false otherwise.

Examples

#[macro_use] extern crate vexyz_math;
use vexyz_math::*;

assert!(bvec4!(false, true, false, false).any());
assert!(!bvec4!(false, false, false, false).any());

fn not(&self) -> Vec4b

Performs component-wise negation of the vector, returning a new vector.

Examples

#[macro_use] extern crate vexyz_math;
use vexyz_math::*;

let u = bvec4!(true, false, true, false);
assert_eq!(u.not(), bvec4!(false, true, false, true));